requirements.txt
-i https://pypi.org/simple
asgiref==3.10.0; python_version >= '3.9'
django==5.2.7; python_version >= '3.10'
django-cors-headers==4.9.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
psycopg2-binary==2.9.11; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
